adding mask creation utility
